# docs/public/api/swagger-2.0.yaml
basePath: /api
definitions:
...
  services.AICaptureItem:
    properties:
      captureGroupId:
        type: string
        x-nullable: true
        x-omitempty: true
      clientId:
        type: string
      description:
        type: string
      entityTypeId:
        type: string
      manufacturer:
        type: string
      modelNumber:
        type: string
      moveDisposition:
        type: string
      moveDispositionNote:
        type: string
      name:
        type: string
      needsReview:
        type: boolean
      photoIds:
        items:
          type: string
        type: array
      photoIndexes:
        items:
          type: integer
        type: array
      quantity:
        type: number
      reviewReason:
        type: string
      tagIds:
        items:
          type: string
        type: array
    type: object
...
  services.AIItemReanalysisOut:
    properties:
      item:
        $ref: '#/definitions/services.AICaptureItem'
      photoCount:
        type: integer
      provider:
        type: string
      warnings:
        items:
          type: string
        type: array
    type: object
...
  v1.aiItemReanalysisRequest:
    properties:
      instruction:
        type: string
    type: object
...
  validate.ErrorResponse:
    properties:
      error:
        type: string
      fields:
        type: string
    type: object
info:
  contact:
    name: Homebox Team
    url: https://discord.homebox.software
  description: Track, Manage, and Organize your Things.
  title: Homebox API
  version: "1.0"
paths:
...
  /v1/ai/items/{id}/reanalyze:
    post:
      consumes:
      - application/json
      parameters:
      - description: Item ID
        in: path
        name: id
        required: true
        type: string
      - description: Optional visual-analysis guidance
        in: body
        name: payload
        schema:
          $ref: '#/definitions/v1.aiItemReanalysisRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/services.AIItemReanalysisOut'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/validate.ErrorResponse'
        "502":
          description: Bad Gateway
          schema:
            $ref: '#/definitions/validate.ErrorResponse'
      security:
      - Bearer: []
      summary: Preview Qwen reanalysis for an existing inventory item
      tags:
      - AI Capture
...
schemes:
- https
- http
securityDefinitions:
  Bearer:
    description: '"Type ''Bearer TOKEN'' to correctly set the API Key"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"